Move code from trait to action
